Add files via upload
This a first aproach using Pyhton 3.8 to the undertanding and processing of a dataset downloaded from Kaggle.com, where based on multiple variables it's necessary to adjust a model to predict with the best accuracy both the nature of the genetic disorder and the disorder subclass. Here, in the first place, I use the decision tree algorithm from the sklearn library.