Add files via upload
Global warming is a major concern because it threatens the carbon fixation and respiration rates of tropical forest, which are very important to the terrestrial net primary production. The SlotWinter2017 dataset provides a series of information of mesurements for multiple variables related to the photosyntetic metabolism of multiple especies of two tropical forest located in Panamá. The goal of the study is to check the impact of temperature on the photosyntetic response.